prediction_ml_framework-ZT/requirements/requirements.txt
torch
torchvision
torchaudio
pytorch_lightning
joblib
lmdb
matplotlib
numpy
onnx
protobuf
waymo_open_dataset_tf_2_6_0
tqdm
onnx_simplifier
PyYAML
pyyaml-include
scipy
